refactor: replace hooks into navgiation folder
