+++ src/hooks/index.ts
@@ -2,7 +2,5 @@ export {default as useFocus} from './useFocus';
 export {default as useWindowBreakpoint} from './useWindowBreakpoint';
 export {default as useMount} from './useMount';
 export {default as useHeightCalculator} from './useHeightCalculator';
-export {default as useActiveNavItem} from './useActiveNavItem';
-export {default as useShowBorder} from './useShowBorder';
 export * from './useAnalytics';
 export * from './hubspot';
